+++ changelog.md
@@ -104,6 +104,8 @@ Codion Change Log
 - EntityPanel, TabbedDetailLayout, now possible to configure the available panel states.
 - EntityComponents now final, constructor private, factory method added.
 - EntityComponents parameter removed from EntityEditComponentPanel constructor.
+### is.codion.swing.framework.ui.tools
+- DomainGeneratorPanel.DEFAULT_USERNAME configuration value added.
 
 ## 0.17.43
 ### is.codion.common.core